Change create() functions
The previous create() function built an anchored URL (e.g.
".../index.php/resource/id?...") and returned it as a string; this
function has been renamed createAnchoredUrl(). The new create() function
builds a URL from a given relative URL; much like a browser would, if
the URL begins with a forward slash it is appended to a site URL,
otherwise it is appended to the directory URL. This new version of
create() helps standardize URLs for CSS, images, etc., from within
pocket-knife.